+++ apps/alert_processor/lib/helpers/date_time_helper.ex
@@ -231,11 +231,4 @@ defmodule AlertProcessor.Helpers.DateTimeHelper do
       _ -> :error
     end
   end
-
-  @spec naive_to_local(NaiveDateTime.t) :: DateTime.t
-  def naive_to_local(naive_datetime) do
-    naive_datetime
-    |> DateTime.from_naive!("Etc/UTC")
-    |> Calendar.DateTime.shift_zone!(@time_zone)
-  end
 end
